Create index.html
Following completed:
- Basic building blocks for navbar.
- Cover titles.
- 2 live issues sections (1 for the static intro part and the other for the three listed live issues).
- Footer (Not added placeholders for their allies which is something we need to ask for too).

The reason why I haven't yet started on the Albany bicentenary, Why History Matters and the 2024 Annual lecture is because I believe Thisari and Zarah mentioned how they are combining both of their concepts to make the whole website consistent after the meeting with Elizabeth and Paul earlier today.

Will be moving onto fleshing out the design and tweaks for the navigation bar and referencing the fonts that was used in the design compositions (Aktiv Grotesk and Adobe Casion Pro).