Please read Chet
If you wanted to replace an image, make sure to just to replace the original image in the folder and not make a duplicate. This only gives you more unnecessary work in editing code that no longer needs editing in the html and confusion with many duplicated images that doesn't need to be in the folder. If not sure what I'm referring to, please contact.